Remove collection of iterators, use more iterators
This makes it easier to later extend the functionality, so that
multiple bits in the pixel can be used and not only the last one.
I would only need to modify the iterator for position, making it also
describe the position in the byte from the pixel.
Since iterators are not collected, it should also be more efficient.